requirements.txt
asgiref==3.2.7
astroid==2.4.1
certifi==2020.4.5.1
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
cycler==0.10.0
Django==3.0.6
django-tastypie==0.14.3
dnspython==2.0.0
docutils==0.16
Flask==1.1.2
Flask-CLI==0.4.0
Flask-Cors==3.0.9
Flask-Ext==0.1
Flask-MySQL==1.5.1
Flask-MySQLdb==0.2.0
Flask-SocketIO==4.3.2
Flask-SQLAlchemy==2.4.4
idna==2.6
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
marshmallow==3.9.1
marshmallow-sqlalchemy==0.24.1
matplotlib==3.3.3
mccabe==0.6.1
mysql==0.0.2
mysqlclient==2.0.1
numpy==1.19.4
Pillow==8.1.0
pycparser==2.20
pylint==2.5.2
pymongo==3.11.2
PyMySQL==0.10.1
pyparsing==2.4.7
pyrh==2.0
python-dateutil==2.8.1
python-engineio==3.14.2
python-mimeparse==1.6.0
python-socketio==4.6.1
pytz==2020.1
redis==3.5.3
requests==2.23.0
Robinhood==1.0.1
six==1.14.0
SQLAlchemy==1.3.20
SQLAlchemy-serializer==1.3.4.4
sqlparse==0.3.1
statistics==1.0.3.5
toml==0.10.1
urllib3==1.22
Werkzeug==1.0.1
wordcloud==1.8.1
wrapt==1.12.1
